+++ pydantic/experimental/pipeline.py
@@ -10,9 +10,10 @@ from collections.abc import Container
 from dataclasses import dataclass
 from decimal import Decimal
 from functools import cached_property, partial
-from typing import TYPE_CHECKING, Annotated, Any, Callable, Generic, Protocol, TypeVar, Union, overload
+from typing import TYPE_CHECKING, Any, Callable, Generic, Protocol, TypeVar, Union, overload
 
 import annotated_types
+from typing_extensions import Annotated
 
 if TYPE_CHECKING:
     from pydantic_core import core_schema as cs
